// Shelf.WebApp/ShelfApp/src/app/shared/models/planogram/planogram.ts
export interface PackageAttributes {
    IdPackageAttribute: number;
    IdCorp: number;
    IdPog?: number;
    IdProduct: number;
    IdPackage: number;
    UniqueProdPkg: string;
    Cost?: number;
    CostInventory?: number;
    Retail?: number;
    Profit?: number;
    DOS?: number;
    DPP?: number;
    Sales?: number;
    SPM?: number;
    AdjustmentFactor: number;
    Movement?: number;
    Turns?: number;
    GMROII?: number;
    ROII?: number;
    ProfitPerUnit?: number;
    LostSales?: number;
    OOS?: number;
    ActualMargin?: number;
    ContributedMargin?: number;
    AdjustedMovement?: number;
    WeeksOfSupply?: number;
    Casepack?: number;
    AdjMovtFactor: number;
    Color: string;
    RecADRI: string;
    RecAppliedOn?: any;
    RecROSUnits?: number;
    RecROSSales?: number;
    RecROSVolume?: number;
    RecCPI?: number;
    RecLoyaltyIndex?: number;
    RecFacings?: number;
    RecSalesFCST?: number;
    RecMovtFCST?: number;
    RecUserNotes: string
    RecSystemNotes: string
    RecNPI?: boolean;
    RecMustStock?: boolean;
    RecMustNotStock?: boolean;
    RecAvailableToOrder?: boolean;
    RecMktUnits?: number;
    RecMktSales?: number;
    RecMktVolume?: number;
    RecVolume?: number;
    RecMarginCash?: number;
    RecMarginPercentage?: number;
    RecProfitCash?: number;
    RecProfitPercentage?: number;
    RecDistNumeric?: number;
    RecDistWeighted?: number;
    RecNoOfStores?: number;
    RecPSW?: number;
    RecCapacity: number;
    RecCases: number;
    RecCostInventory: number;
    RecDOS: number;
    RecFrontsDeep: number;
    RecROII: number;
    RecTurns: number;
    RecWOS: number;
    CurrNoOfStores?: number;
    CurrPSW?: number;
    CurrMovt?: number;
    CurrMovtLY?: number;
    CurrMovtAdj?: number;
    CurrSales?: number;
    CurrSalesLY?: number;
    CurrGrossProfit?: number;
    CurrAdjGrossProfit?: number;
    CurrTrueProfit?: number;
    ManualPerfData: string;
    ForceFit: boolean
    WhiteSpacePosition: number;
    WhiteSpaceWidth: number;
    WhiteSpaceText: string;
    FacingsY?: number
}